% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Numeric Column Summary}
\usage{
numeric_summary(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with summary statistics for each numeric column.
}
\description{
Summarizes numeric columns in a data frame with common descriptive
statistics.
}
